Buff Handler Refactor and Buffs class renamed to BuffCollection
* Buffs class and Buffs folder were causing namespace issues.
* Buff Handler is called from the Buff class itself instead of Buffs (BuffCollection).

+++ src/ChannelServer/Network/Helpers/BuffHelper.cs
@@ -16,7 +16,7 @@ namespace Melia.Channel.Network.Helpers
 			packet.PutInt(buff.Target.Handle);
 			packet.PutInt(buff.Handle);
 			packet.PutInt((int)buff.Id);
-			packet.PutInt(buff.SkillId);
+			packet.PutInt((int)buff.SkillId);
 			packet.PutInt(0);
 			packet.PutInt(0);
 			packet.PutInt(0);
